requirements.txt
pandas
numpy
matplotlib
seaborn
openpyxl
python-docx
pytesseract
PyMuPDF
together
requests
pillow
streamlit
